requirements.txt
kafka-python
Flask
csv_to_json
elasticsearch
requests